libavfilter/vf_objectdetect: add a filter for object detection
This filter is not finished yet. Currently, for visual effect, it
detects objections and draws a box for the detected objections with score > 0.8.

The purpose here is to show the previous changes of the dnn is necessary,
and it is also a RFC (require for comment) patch for this filter.

In my plan, an example under doc/examples will be added to show how to
use this filter and connet the filter result with the ROI encoding.

And, will add options to this filter for score_threshold, visual effect or not.

And, a new type of sidedata will be added to hold the filter result.
So, the example can get the filter result and set ROI info for encoder.

Looks that, it is not easy to transfer data between filters, between
filter and application, between filter and encoders, the current feasible
method I found is to by sidedata.