lavc/vaapi_encode: remove consistency check of depth for each component
Previously, consistency check of depth is introduced to make sure
each component of the pixel format has the same depth.

Since Y410(with a different depth in alpha channel) is supported for
ICL+, this check should be removed to make the encoder work.

Not quite proper since comp[0].depth may not equal to or larger than the
depth in other component.

Signed-off-by: Linjie Fu <linjie.fu@intel.com>