lavc/vaapi_encode: Async the encoding and output procedure of encoder
Currently, vaapi encodes a pic if all its references are ready,
and then outputs it immediately by calling vaapi_encode_output.

However, while working on output procedure, hardware is be able to
cope with encoding tasks in the meantime to have the better performance.

So a more efficient way is to send all the pics with available refs to
hardware to allow encoding while output.

It's what vaapi originally did before the regression, and the performance
could be improved for ~20%.

CMD:
ffmpeg -hwaccel vaapi -vaapi_device /dev/dri/renderD128
-hwaccel_output_format vaapi -i bbb_sunflower_1080p_30fps_normal.mp4
-c:v h264_vaapi -f h264 -y /dev/null

Source:
https://download.blender.org/demo/movies/BBB/

Before:
    ~164 fps
After:
    ~198 fps

Fix #7706.

Signed-off-by: Linjie Fu <linjie.fu@intel.com>