avutil/frame: add use_last_roi
For some cases, the regions of interest do not change, it is not
convenient to always prepare the roi data for every frame. So, add
use_last_roi to show it uses the same roi data as last frame.

Since a new flag is added into AVFrame, the major version number of
lavu is changed.