qsv: add ${includedir}/mfx to the search path for old versions of libmfx
${includedir}/mfx has been added to Cflags in libmfx.pc for the current
libmfx. We may add ${includedir}/mfx to the search path for olda
versions of libmfx so that we may include foo.h instead of mfx/foo.h

After applying this change, we won't need to change #include to include
the right header files when the mfx header files are installed to a new
directory for a new version of libmfx. E.g. for libmfx 2.0 (oneVPL,
https://github.com/oneapi-src/oneVPL), the header files will be
installed in vpl directory

If your libmfx comes without pkg-config support, this patch requires a
small change to your environment setting (e.g. adding
/opt/intel/mediasdk/include/mfx instead of /opt/intel/mediasdk/include
to CFLAGS) so that the build can find the headers