libavcodec/qsvdec: fix bug: pre-initial twice.
When program first open a video, it will call qsv_decode_preinit()
twice. Even if the resource is released in second call, it still takes
more device memory because of the delay between calling release api and
really releasing memory in device.
Now I add a flag to control the program so that it only call qsv_decode_preinit()
once when it first open a video.

Signed-off-by: Wenbin Chen <wenbin.chen@intel.com>