fftools/ffmpeg_qsv.c: reduce initial_pool_size to reduce device memory use
ffmpeg-qsv prealloc at least 64 frames in device memory. It is too
much because the most of them are useless. Now I reduce this size to 32,
taking the configuration of "examples/qsvdec.c" as a reference. If the
frame pool is not enough, user can use "-extra_hw_frames" parameter to
extend it.

Signed-off-by: Wenbin Chen <wenbin.chen@intel.com>