libavutil/hwcontext_qsv.c:Fill pixels for realigned frame in qsv
According to MediaSDK doc, app needs to fill the extend area caused by alignment,
So I fill the copy of pixel on the border.
MediaSDK doc: https://github.com/Intel-Media-SDK/MediaSDK/blob/master/doc/mediasdk-man.md#encoding-procedures,
"It is the application's responsibility to fill pixels outside of crop window when it is smaller than frame to be encoded.
Especially in cases when crops are not aligned to minimum coding block size (16 for AVC, 8 for HEVC and VP9)."

Signed-off-by: Zhang Yuankun <yuankunx.zhang@intel.com>