libavfilter/qsvvpp: change the output frame's width and height
qsvvpp align the width and height with 16, and that may lead to error.
For example, when we use qsvvpp to resize frame to 1080p, qsvvpp will
align frame to 1088 which is different from the height of
encoder (1080). Now I assign the out_link's width and height to output's
frame to overwrite the hw_frame_ctx's resolution.

Signed-off-by: Wenbin Chen <wenbin.chen@intel.com>