libavutil/hwcontext_qsv: Change width and height of hw_frame_ctx to original value
The AVFrame created from AVHWFrameContext should have original w/h,
otherwise it is confusing about the real resolution of frame, so I
changed the w/h of hw_frame_ctx to original value.

The w/h will be aligned when we need aligned number, for example, when
we create vaSurface and when we configure mfxSurface. The aligned number
for different codecs are different, so I add two variables into
AVQSVFrameContext to store aligned w/h.

Because I change the w/h of hw_frame_ctx, the avframe used to hwdownload
and hwupload will change too. The qsvvpp need aligned space to store
data, and the AVFrame may do not have sufficient space so I add realloc
operation into hwdownload and hwupload.

Signed-off-by: Wenbin Chen <wenbin.chen@intel.com>