libavutil/hwcontext_vulkan: Add single_linear flag to make vulkan compatible with vaapi device.
Vaapi can import external surface, but all the planes of the external
frames should be in the same drm object. A new flag is introduced and
vulkan can choose to allocate planes in one memory according this flag.
This flag will be enabled when the vulkan device is derived from vaapi
device, so that this change will not affect current vulkan behaviour.

Signed-off-by: Wenbin,Chen <wenbin.chen@intel.com>