libavfilter/x86/vf_gblur: add localbuf and ff_horiz_slice_avx2/512()
We introduced a new ff_horiz_slice_avx2() and ff_horiz_slice_avx512()
implemented on a new algorithm. In a nutshell, the new algorithm does three
things, gathering data from 8/16 rows, blurring data, and scattering data
back. Here we used a customized transpose8x8/16x16 to avoid the huge overhead
brought by scattering, which is dependent on the temporary buffer called
localbuf added newly.

The fps of the old horiz_slice SIMD was 72.5 approximately, and the fps
of the new horiz_slice_avx2/512 could reach 416~500. The actual performance
would depend on whether the height of the sliced frame is aligned with
the 8/16.

Co-authored-by: Cheng Yanfei <yanfei.cheng@intel.com>
Co-authored-by: Jin Jun <jun.i.jin@intel.com>
Signed-off-by: Wu Jianhua <jianhua.wu@intel.com>