avfilter/avfiltergraph: add an auto hwmap filter
When two formats lists cannot be merged, a scale filter is
auto-inserted. However, when it comes to hardware map, we have to
manually add a hwmap filter to do the conversion. This patch introduces
an auto hwmap filter to do the hwmap conversion automatically and
modifies the negotiation structures, making it extensible for more auto
filters.

Signed-off-by: Tong Wu <tong1.wu@intel.com>