libavcodec/mpegvideo_enc: Fix a chroma mb size error in sse_mb()
For 422 frames we should not use hard coded 8 to calculate mb size for
uv plane. Chroma shift should be taken into consideration to be
compatiple with different sampling format.

The error occurs when av_cpu_max_align() return 64 on the platform
supporting AVX512. This is a hidden error and it is exposed after commit
17a59a634c39b00a680c6ebbaea58db95594d13d.

Signed-off-by: Wenbin Chen <wenbin.chen@intel.com>