lavc/qsvenc: use VBR by default on Windows
Currently AVBR is disabled and VBR is the default method on Linux, but
AVBR is the default one on Windows. In order to make user experience
better accorss Linux and Windows, use VBR by default too on Windows.
User need to set avbr_accuracy and avbr_convergence to non-zero values
explicitly if AVBR is expected.

In addition, AVBR works for H264 and HEVC only in the SDK.

$ ffmpeg.exe -v verbose -f lavfi -i yuvtestsrc -vf "format=nv12" -c:v
vp9_qsv -f null -